sprintf format fixes for 64bit systems.

+++ trunk/src/i_complete.cpp
@@ -279,7 +279,7 @@ void ExComplete::RepaintStatus()
 	p += l;
 	char s[100];
 	sprintf(s, "] (T:%d/%d  S:%d)", WordPos + 1, WordsLast,
-		WordFixedCount);
+		(int)WordFixedCount);
 	MoveStr(B, p, W, s, COM_NORM, W);
 	// ConSetCursorPos(cur + WordFixed, H - 1);
     } else
